# obsrvr-lake/stellar-query-api/go/docs/swagger.json
{
    "basePath": "/",
    "definitions": {
...
        "main.SmartWalletInfo": {
            "properties": {
                "contract_id": {
                    "type": "string"
                },
                "has_check_auth": {
                    "type": "boolean"
                },
                "is_smart_wallet": {
                    "type": "boolean"
                },
                "signer_count": {
                    "type": "integer"
                },
                "signers": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                }
            },
            "type": "object"
        },
...
        "main.TxBalanceChange": {
            "properties": {
                "address": {
                    "type": "string"
                },
                "after": {
                    "type": "string"
                },
                "asset_code": {
                    "type": "string"
                },
                "asset_issuer": {
                    "type": "string"
                },
                "asset_type": {
                    "type": "string"
                },
                "before": {
                    "type": "string"
                },
                "delta": {
                    "type": "string"
                }
            },
            "type": "object"
        },
        "main.TxDiffs": {
            "properties": {
                "balance_changes": {
                    "items": {
                        "$ref": "#/definitions/main.TxBalanceChange"
                    },
                    "type": "array"
                },
                "ledger_sequence": {
                    "type": "integer"
                },
                "state_changes": {
                    "items": {
                        "$ref": "#/definitions/main.TxStateChange"
                    },
                    "type": "array"
                },
                "tx_hash": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "main.TxStateChange": {
            "properties": {
                "after": {
                    "type": "string"
                },
                "before": {
                    "type": "string"
                },
                "entry_type": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
    },
    "host": "gateway.withobsrvr.com",
    "info": {
        "contact": {
            "email": "support@obsrvr.com",
            "name": "OBSRVR Team",
            "url": "https://obsrvr.com"
        },
        "description": "API for querying Stellar blockchain data across Bronze, Silver, Semantic, and Gold data layers.\n\n## Data Layers\n- **Bronze**: Raw blockchain data (ledgers, transactions, operations, effects)\n- **Silver**: Enriched and processed data (accounts, assets, transfers, trades)\n- **Semantic**: Meaning-oriented views (activities, contracts, value flows)\n- **Gold**: Analytical views (snapshots, compliance reports)\n- **Index Plane**: Fast lookup services (transaction hash, contract events)\n\n## Amount Formatting\nAll amounts are returned in decimal format with 7 decimal places (Stellar stroops conversion).\nExample: Raw stroops `10000000` is formatted as `\"1.0000000\"`",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "title": "Stellar Query API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/silver/smart-wallet/{contract_id}": {
            "get": {
                "description": "Heuristically detects if a contract implements SEP-50 smart wallet by checking for __check_auth events and signer storage",
                "parameters": [
                    {
                        "description": "Contract ID (C...)",
                        "in": "path",
                        "name": "contract_id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Smart wallet info",
                        "schema": {
                            "$ref": "#/definitions/main.SmartWalletInfo"
                        }
                    },
                    "400": {
                        "description": "Missing contract_id",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Detect SEP-50 smart wallet",
                "tags": [
                    "Smart Wallets"
                ]
            }
        },
...
        "/api/v1/silver/tx/{hash}/diffs": {
            "get": {
                "description": "Decodes tx_meta XDR to extract balance changes and state changes for a transaction",
                "parameters": [
                    {
                        "description": "Transaction hash",
                        "in": "path",
                        "name": "hash",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Transaction diffs",
                        "schema": {
                            "$ref": "#/definitions/main.TxDiffs"
                        }
                    },
                    "400": {
                        "description": "Missing transaction hash",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Get transaction balance/state diffs",
                "tags": [
                    "Transactions"
                ]
            }
        },
...
    },
    "schemes": [
        "https"
    ],
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication. Format: \"Api-Key YOUR_API_KEY\"",
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}